backend/requirements.txt
asgiref==3.5.2
boto3==1.24.70
botocore==1.27.70
certifi==2022.6.15.1
charset-normalizer==2.1.1
dj-stripe==2.6.2
Django==4.1.1
django-anymail==8.6
django-cors-headers==3.13.0
django-storages==1.13.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
idna==3.3
jmespath==1.0.1
jsonfield==3.1.0
Pillow==9.2.0
psycopg2==2.9.3
PyJWT==2.4.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2022.2.1
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
stripe==4.1.0
tzdata==2022.2
urllib3==1.26.12
whitenoise==6.2.0
